Add MapUtil and make significant changes to Cart logic

+++ src/main/java/com/farid/ahadli/my_restaurant_app/repository/RestaurantIngredientsRepository.java
@@ -3,6 +3,6 @@ package com.farid.ahadli.my_restaurant_app.repository;
 import com.farid.ahadli.my_restaurant_app.model.entity.RestaurantIngredients;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RestaurantIngredientsRepository extends JpaRepository<RestaurantIngredients, Integer> {
+public interface RestaurantIngredientsRepository extends JpaRepository<RestaurantIngredients, Long> {
 
 }
