+++ src/main/java/com/farid/ahadli/my_restaurant_app/repository/RestaurantOrdersRepository.java
@@ -3,6 +3,6 @@ package com.farid.ahadli.my_restaurant_app.repository;
 import com.farid.ahadli.my_restaurant_app.model.entity.RestaurantOrders;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RestaurantOrdersRepository extends JpaRepository<RestaurantOrders, Integer> {
+public interface RestaurantOrdersRepository extends JpaRepository<RestaurantOrders, Long> {
 
 }
